move automated checks view (the container) on level up
