+++ src/tests/unit/tests/electron/views/automated-checks/automated-checks-view.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { AutomatedChecksView, AutomatedChecksViewProps } from 'electron/views/automated-checks/components/automated-checks-view';
+import { AutomatedChecksView, AutomatedChecksViewProps } from 'electron/views/automated-checks/automated-checks-view';
 import { shallow } from 'enzyme';
 import * as React from 'react';
